[borders] split x11 and wayland injections
Prevents import errors for x11 users as code tried to import wayland
modules.
