+++ CHANGELOG
@@ -1,3 +1,4 @@
+2024-07-16: [BUGFIX] Fix bug where border decorations failed on x11 if user did not have wayland libraries installed
 2024-07-12: [RELEASE] v0.27.0 release - compatible with qtile 0.27.0
 2024-07-12: [BUGFIX] Fix bug where conditional border could resize window
 2024-07-07: [FEATURE] Add `RoundedCorners` border decoration (Wayland only)
